Make with_new_array_scope easier to use
First, you couldn’t actually use it because you had to supply
F and R if you wanted to supply N, which you had to supply.

Second, we now compute the ScopedHandle object for you,
so you no longer have to Scope::get them (which returns Option).
And the compiler will catch out-of-bounds access if you use constant indices.